+++ docs/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+* Percentage from total time was outside of borders (to the right)
+
 
 ## 1.1.0 - 2018-11-12
 ### Add