+++ docs/CHANGELOG.md
@@ -5,11 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## 1.2.0-alpha - 2018-11-19
 ### Added
 * Press and hold stop button: stop and discard the work-session (with a confirmation box)
 
 ### Changed
-* Dashboard summary styles
+* Improve dashboard summary styles
 
 ### Fixed
 * Percentage from total time was outside of borders (to the right)